Protecting API Routes by authenticate Middleware

+++ server.js
@@ -2,6 +2,7 @@ const express = require('express')
 const cors = require('cors')
 const app = express()
 const {mongoose} = require('./db/mongoose')
+const jwt = require('jsonwebtoken')
 
 const { List, Task, User } = require('./db/models')
 
@@ -26,6 +27,21 @@ const corsOptions = {
 };
 
 app.use(cors(corsOptions));
+
+/* Check if the request hase a valid JWT access token */
+let authenticate = (req, res , next) => {
+    let accessToken = req.header('x-access-token')
+
+    jwt.verify(accessToken, User.getJWTSecret(), (err, decoded) => {
+        if(err) {
+            // JWT invalid - DO NOT AUTHENTICATE
+            res.status(401).send(err)
+        } else {
+            req.user_id = decoded.id
+            next()
+        }
+    })
+}
 
 /* Verify Refresh Token Middleware */
 let verifySession = (req, res, next) => {
@@ -66,10 +82,12 @@ let verifySession = (req, res, next) => {
 
 /**
  * GET /lists
- * Purpose: Get all lists
+ * Purpose: Get all lists that belong to the authenticated user
  */
-app.get('/lists', (req, res) => {
-    List.find({}).then((lists) => {
+app.get('/lists', authenticate, (req, res) => {
+    List.find({
+        _userId: req.user_id
+    }).then((lists) => {
         res.send(lists)
     })
 })
@@ -78,9 +96,10 @@ app.get('/lists', (req, res) => {
  * POST /lists
  * Purpose: Create a list
  */
-app.post('/lists', (req, res) => {
+app.post('/lists', authenticate, (req, res) => {
     let newList = new List({
-        title: req.body.title
+        title: req.body.title,
+        _userId: req.user_id
     })
     newList.save().then((list) => {
         res.send(list)
@@ -91,8 +110,8 @@ app.post('/lists', (req, res) => {
  * PATCH /lists/:id
  * Purpose: Update a specified list
  */
-app.patch('/lists/:id', (req, res) => {
-    List.findByIdAndUpdate({_id: req.params.id}, {
+app.patch('/lists/:id', authenticate, (req, res) => {
+    List.findByIdAndUpdate({_id: req.params.id, _userId: req.user_id}, {
         $set: req.body
     }).then(() => {
         res.sendStatus(200)
@@ -100,14 +119,16 @@ app.patch('/lists/:id', (req, res) => {
 })
 
 /**
- * PATCH /lists/:id
- * Purpose: Delete a specified list
+ * DELETE /lists/:id
+ * Purpose: Delete a specified list and all the tasks that are in the list
  */
-app.delete('/lists/:id', (req, res) => {
+app.delete('/lists/:id', authenticate, (req, res) => {
     List.findByIdAndRemove({
-        _id: req.params.id
+        _id: req.params.id,
+        _userId: req.user_id
     }).then((removedList) => {
         res.send(removedList)
+        deleteTaskFromList(removedList._id)
     })
 })
 
@@ -115,11 +136,26 @@ app.delete('/lists/:id', (req, res) => {
  * GET /lists/:listId/tasks
  * Purpose: Get all tasks that belong to a specified list
  */
- app.get('/lists/:listId/tasks', (req, res) => {
-    Task.find({
-        _listId: req.params.listId
-    }).then((tasks) => {
-        res.send(tasks)
+ app.get('/lists/:listId/tasks', authenticate, (req, res) => {
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            // User authenticated because a list was found so he can view the tasks
+            return true
+        }
+        return false
+    }).then((canViewTasks) => {
+        if (canViewTasks) {
+            Task.find({
+                _listId: req.params.listId
+            }).then((tasks) => {
+                res.send(tasks)
+            })
+        } else {
+            res.sendStatus(404)
+        }
     })
 })
 
@@ -127,12 +163,27 @@ app.delete('/lists/:id', (req, res) => {
  * GET /lists/:listId/tasks/:taskId
  * Purpose: Get all tasks that belong to a specified list
  */
- app.get('/lists/:listId/tasks/:taskId', (req, res) => {
-    Task.findOne({
-        _id: req.params.taskId,
-        _listId: req.params.listId
-    }).then((tasks) => {
-        res.send(tasks)
+ app.get('/lists/:listId/tasks/:taskId', authenticate, (req, res) => {
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            // User authenticated because a list was found so he can view the specified task
+            return true
+        }
+        return false
+    }).then((canViewTask) => {
+        if (canViewTask) {
+            Task.findOne({
+                _id: req.params.taskId,
+                _listId: req.params.listId
+            }).then((tasks) => {
+                res.send(tasks)
+            })
+        } else {
+            res.sendStatus(404)
+        }
     })
 })
 
@@ -140,13 +191,28 @@ app.delete('/lists/:id', (req, res) => {
  * POST /lists/:listId/tasks
  * Purpose: Create a task in a specified list
  */
- app.post('/lists/:listId/tasks', (req, res) => {
-    let newTask = new Task({
-        title: req.body.title,
-        _listId: req.params.listId,
-    })
-    newTask.save().then((task) => {
-        res.send(task)
+ app.post('/lists/:listId/tasks', authenticate, (req, res) => {
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            // User authenticated because a list was found so he can create a new task
+            return true
+        }
+        return false
+    }).then((canCreateTask) => {
+        if (canCreateTask) {
+            let newTask = new Task({
+                title: req.body.title,
+                _listId: req.params.listId,
+            })
+            newTask.save().then((task) => {
+                res.send(task)
+            })
+        } else {
+            res.sendStatus(404)
+        }
     })
 })
 
@@ -154,28 +220,59 @@ app.delete('/lists/:id', (req, res) => {
  * PATCH /lists/:listId/tasks/:taskId
  * Purpose: Update an existing task
  */
- app.patch('/lists/:listId/tasks/:taskId', (req, res) => {
-    Task.findByIdAndUpdate({
-        _id: req.params.taskId,
-        _listId: req.params.listId
-    }, {
-        $set: req.body
-    }).then(() => {
-        res.sendStatus(200)
-    })
+ app.patch('/lists/:listId/tasks/:taskId', authenticate, (req, res) => {
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            // User authenticated so he can update the specified task
+            return true
+        }
+        return false
+    }).then((canUpdateTask) => {
+        if(canUpdateTask) {
+            Task.findByIdAndUpdate({
+                _id: req.params.taskId,
+                _listId: req.params.listId
+            }, {
+                $set: req.body
+            }).then(() => {
+                res.sendStatus(200)
+            })
+        } else {
+            res.sendStatus(404)
+        }
+    })    
 })
 
 /**
  * DELETE /lists/:listId/tasks/:taskId
  * Purpose: Delete a specified list
  */
- app.delete('/lists/:listId/tasks/:taskId', (req, res) => {
-    Task.findByIdAndRemove({
-        _id: req.params.taskId,
-        _listId: req.params.listId
-    }).then((removedList) => {
-        res.send(removedList)
+ app.delete('/lists/:listId/tasks/:taskId', authenticate, (req, res) => {
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            // User authenticated so he can remove the specified task
+            return true
+        }
+        return false
+    }).then((canDeleteTask) => {
+        if (canDeleteTask) {
+            Task.findByIdAndRemove({
+                _id: req.params.taskId,
+                _listId: req.params.listId
+            }).then((removedList) => {
+                res.send(removedList)
+            })
+        } else {
+            res.sendStatus(404)
+        }
     })
+    
 })
 
 /**
@@ -242,6 +339,13 @@ app.get('/users/me/access-token', verifySession, (req, res) => {
     res.status(400).send(err)
    })
 })
+
+/* Helper Methods */
+let deleteTaskFromList = (_listId) => {
+    Task.deleteMany({
+        _listId
+    })
+}
 
 app.listen(3000, () => {
     console.log("Server is listening on port 3000");
